requirement.txt
pyspark==3.0.1
azure-core==1.10.0
azure-eventhub==5.3.0
azure-eventhub-checkpointstoreblob-aio==1.1.2
azure-storage-blob==12.7.1
pymongo==3.11.0
dnspython==1.16.0
PyYAML
pandas
-e .